requirements.txt
networkx
pyyaml
matplotlib